Replace manual SQL with ORM
Introduced 'peewee' as ORM and modified code to not execute hand-crafted
SQL anymore, but make use of ORM to store and fetch data.

Most probably some fine tuning needs to be done, but overall it seems to
work and be a pretty good replacement with hopefully less hassle to
enhance it. Except there's no migration strategy yet, in case of changes
on data models, but that's not worse than before.